requirements.txt
Django==4.0.2
djangorestframework==3.13.1
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.14.0
